Clean up the morale and fatigue code after review
Reuse and duplication:
- Drop the second per-tick fat_nocharge vector in the damage apply
  pass; the one built at the top of step_sim is owned data and is
  still live there.
- Fuse the routing-ally and routing-enemy scans into a single pass
  over the broken list instead of two filters differing only by team.
- Read FL_SEED as an integer through the existing env_or helper
  rather than parsing a float and casting.

Simplification:
- Collapse the four parallel matches over the fatigue bands into one
  const table indexed by state, with the accessors as thin lookups.
- Delete MoraleFactors::effective and ::friends, which had no readers.
- Replace the cached shaken and wavering flags with a derived band()
  helper next to morale01, removing two GroupData fields and the
  hand-maintained obligation to clear them in every other state arm.

Efficiency:
- Compare squared distances in the per-regiment neighbour scans, which
  run for every regiment every tick; this matches the convention the
  regiment-pair loop in frontline.rs already uses.
- Reuse the broken and steady centroid snapshots across ticks through
  Local buffers instead of allocating both every tick.

+++ src/units.rs
@@ -126,7 +126,7 @@ pub fn hash01(mut x: u32) -> f32 {
 /// so run-to-run variance must be injected at spawn).
 fn spawn_seed_offset() -> u32 {
     static S: std::sync::OnceLock<u32> = std::sync::OnceLock::new();
-    *S.get_or_init(|| (crate::util::env_or("FL_SEED", 0.0) as u32).wrapping_mul(0x9E37_79B1))
+    *S.get_or_init(|| crate::util::env_or("FL_SEED", 0u32).wrapping_mul(0x9E37_79B1))
 }
 
 /// Append one fully-initialized unit to every SoA column. `anchor` is the
